Crashes fixed and test passes; paired w/@brennie (thanks!).  But...
We probably don't want to go with this exact approach, because the NimbusClient participates in cycle collection,
and objects that do that can't implement a threadsafe nsISupports.  This patch really just hacks around it.
Barret has suggested another possible approach, and a few others have come up during other conversations as well...